+++ components/debug-views/r3f.ts
@@ -1,4 +1,10 @@
 export { DebugViewLayer, type DebugViewLayerProps } from "./debug-view-layer"
+export {
+  DebugViewLeva,
+  DEBUG_VIEW_LEVA_THEME,
+  mountDebugViewLeva,
+  type DebugViewLevaProps,
+} from "./debug-view-leva"
 export { DebugViews, type DebugViewsProps } from "./debug-views-post"
 export {
   useDebugViewsControls,